CBT_Logic/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.8.4
aiolimiter==1.2.1
aioredis==2.0.1
aiosignal==1.3.2
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anthropic==0.8.1
anyio==4.8.0
anytree==2.12.1
appnope==0.1.4
arcade==2.6.17
array-record==0.4.1
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
bce-python-sdk==0.9.29
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.1
camel-converter==4.0.1
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dashscope==1.14.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
dill==0.3.9
distlib==0.3.9
distro==1.9.0
Django==4.2.19
dm-tree==0.1.8
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
etils==1.5.2
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.7.4
Farama-Notifications==0.0.4
fastapi==0.115.8
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.4.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.6.0
gitdb==4.0.12
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-auth==2.38.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.67.0rc1
googlesearch-python==1.3.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.63.0rc1
grpcio-tools==1.62.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.25.2
hyperframe==6.1.0
idna==3.10
imageio==2.35.1
imap-tools==1.5.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.27.0
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.7.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.6.0
kiwisolver==1.4.7
lancedb==0.4.0
langchain==0.0.352
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
lazy-object-proxy==1.10.0
lazy_loader==0.4
libclang==18.1.1
libcst==1.0.1
loguru==0.6.0
lxml==5.3.1
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
meilisearch==0.21.0
metagpt @ file:///Users/bohan/Desktop/werewolf_ai_agents
ml-dtypes==0.4.1
more-itertools==10.6.0
motor==3.7.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.24.3
openai==1.64.0
openapi-core==0.18.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.0.3
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pexpect==4.9.0
pika==1.3.2
pillow==11.1.0
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.2
playwright==1.50.0
pluggy==1.5.0
portalocker==2.10.1
prance==23.6.21.0
promise==2.3
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.0
pydantic_core==2.27.2
pyee==12.1.1
pygame==2.6.1
pyglet==2.0.dev23
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.9.0
pymongo==4.11.1
pymunk==6.4.0
pyparsing==3.1.4
pyserial==3.5
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.20
pytiled-parser==2.2.0
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
qdrant-client==1.7.0
qianfan==0.3.2
rank-bm25==0.2.2
ratelimiter==1.2.0.post0
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
semantic-kernel==0.4.3.dev0
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sklearn==0.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
ta==0.10.2
tenacity==8.2.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.5.2
tokenizers==0.20.1
toml==0.10.2
tomli==2.1.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
transformers==4.45.2
typer==0.15.1
typing-inspect==0.8.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultralytics==8.3.27
ultralytics-thop==2.0.10
unet==0.7.7
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.6.2
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.15.0
yarl==1.18.3
zhipuai==2.0.1
zipp==3.20.2
